removed obsolete trait; Session now injected in the bag
Refs #6.0.x
