+++ src/Session/Manager.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace Phalcon\Session;
 
 use Phalcon\Di\Traits\InjectionAwareTrait;
-use Phalcon\Session\Traits\ManagerMagicTraits;
 use SessionHandlerInterface;
 
 use function headers_sent;
@@ -39,7 +38,6 @@ use function session_status;
 class Manager implements ManagerInterface
 {
     use InjectionAwareTrait;
-    use ManagerMagicTraits;
 
     /**
      * @var SessionHandlerInterface|null
@@ -71,6 +69,51 @@ class Manager implements ManagerInterface
     public function __construct(array $options = [])
     {
         $this->setOptions($options);
+    }
+
+    /**
+     * Alias: Gets a session variable from an application context
+     *
+     * @param string $key
+     *
+     * @return mixed
+     */
+    public function __get($key)
+    {
+        return $this->get($key);
+    }
+
+    /**
+     * Alias: Check whether a session variable is set in an application context
+     *
+     * @param string $key
+     *
+     * @return bool
+     */
+    public function __isset(string $key): bool
+    {
+        return $this->has($key);
+    }
+
+    /**
+     * Alias: Sets a session variable in an application context
+     *
+     * @param string $key
+     * @param mixed  $value
+     */
+    public function __set(string $key, $value): void
+    {
+        $this->set($key, $value);
+    }
+
+    /**
+     * Alias: Removes a session variable from an application context
+     *
+     * @param string $key
+     */
+    public function __unset(string $key): void
+    {
+        $this->remove($key);
     }
 
     /**
