+++ tests/integration/Session/Bag/ConstructCest.php
@@ -40,7 +40,7 @@ class ConstructCest
 
         $this->setNewFactoryDefault();
         $this->setDiService('sessionStream');
-        $collection = new Bag('BagTest', $this->container);
+        $collection = new Bag($this->container->get("session"), 'BagTest');
 
         $class = Bag::class;
         $I->assertInstanceOf($class, $collection);
