+++ tests/integration/Session/Bag/CountCest.php
@@ -46,7 +46,7 @@ class CountCest
             'five'  => 'six',
         ];
 
-        $collection = new Bag('BagTest', $this->container);
+        $collection = new Bag($this->container->get("session"), 'BagTest');
 
         $collection->init($data);
