+++ tests/integration/Session/Bag/GetCest.php
@@ -45,7 +45,7 @@ class GetCest
             'three' => 'four',
             'five'  => 'six',
         ];
-        $collection = new Bag('BagTest', $this->container);
+        $collection = new Bag($this->container->get("session"), 'BagTest');
         $collection->init($data);
 
         $expected = 'four';
