+++ tests/integration/Session/Bag/GetIteratorCest.php
@@ -46,7 +46,7 @@ class GetIteratorCest
             'five'  => 'six',
         ];
 
-        $collection = new Bag('BagTest', $this->container);
+        $collection = new Bag($this->container->get("session"), 'BagTest');
         $collection->init($data);
 
         foreach ($collection as $key => $value) {
