+++ tests/integration/Session/Bag/HasCest.php
@@ -46,7 +46,7 @@ class HasCest
             'five'  => 'six',
         ];
 
-        $collection = new Bag('BagTest', $this->container);
+        $collection = new Bag($this->container->get("session"), 'BagTest');
         $collection->init($data);
 
         $actual = $collection->has('three');
