+++ tests/integration/Session/Bag/InjectableCest.php
@@ -46,7 +46,7 @@ class InjectableCest
             'three' => 'four',
             'five'  => 'six',
         ];
-        $collection = new Bag('BagTest', $this->container);
+        $collection = new Bag($this->container->get("session"), 'BagTest');
         $collection->init($data);
 
         /**
