+++ tests/integration/Session/Bag/JsonSerializeCest.php
@@ -46,7 +46,7 @@ class JsonSerializeCest
             'five'  => 'six',
         ];
 
-        $collection = new Bag('BagTest', $this->container);
+        $collection = new Bag($this->container->get("session"), 'BagTest');
 
         $collection->init($data);
         $I->assertEquals($data, $collection->jsonSerialize());
