+++ tests/integration/Session/Bag/ToJsonCest.php
@@ -46,7 +46,7 @@ class ToJsonCest
             'five'  => 'six',
         ];
 
-        $collection = new Bag('BagTest', $this->container);
+        $collection = new Bag($this->container->get("session"), 'BagTest');
 
         $collection->init($data);
         $expected = json_encode($data);
